refactor: rename CompetitionPage to RecentCompetitionPage
