+++ frontend_profile/src/pages/RecentCompetitionPage/components/Competition.tsx
