+++ frontend_profile/src/pages/RecentCompetitionPage/components/CompetitionList.tsx
